renamed plotting/plothandler directory to plotting/_plothandler to remove ambiguity when trying to import plot_handler module
